+++ src/test/debugger/extension/configProvider/provider.unit.test.ts
@@ -17,9 +17,9 @@ import { IFileSystem, IPlatformService } from '../../../../client/common/platfor
 import { IPythonExecutionFactory, IPythonExecutionService } from '../../../../client/common/process/types';
 import { IConfigurationService, ILogger, IPythonSettings } from '../../../../client/common/types';
 import { DebuggerTypeName } from '../../../../client/debugger/constants';
-import { ConfigurationProviderUtils } from '../../../../client/debugger/extension/configProviders/configurationProviderUtils';
-import { PythonV2DebugConfigurationProvider } from '../../../../client/debugger/extension/configProviders/pythonV2Provider';
-import { IConfigurationProviderUtils } from '../../../../client/debugger/extension/configProviders/types';
+import { ConfigurationProviderUtils } from '../../../../client/debugger/extension/configuration/configurationProviderUtils';
+import { PythonV2DebugConfigurationProvider } from '../../../../client/debugger/extension/configuration/pythonV2Provider';
+import { IConfigurationProviderUtils } from '../../../../client/debugger/extension/configuration/types';
 import { DebugOptions, LaunchRequestArguments } from '../../../../client/debugger/types';
 import { IInterpreterHelper } from '../../../../client/interpreter/contracts';
 import { IServiceContainer } from '../../../../client/ioc/types';
